+++ src/Credfeto.Dispatcher.Storage/migrations/004_RemoveDiscordIntegration.sql
@@ -1,0 +1,23 @@
+IF OBJECT_ID(N'[dbo].[NotificationQueue_Upsert]', N'P') IS NOT NULL
+  BEGIN
+    DROP PROCEDURE [dbo].[NotificationQueue_Upsert];
+  END;
+GO
+
+IF OBJECT_ID(N'[dbo].[NotificationQueue_Delete]', N'P') IS NOT NULL
+  BEGIN
+    DROP PROCEDURE [dbo].[NotificationQueue_Delete];
+  END;
+GO
+
+IF OBJECT_ID(N'[dbo].[NotificationQueue_GetReady]', N'P') IS NOT NULL
+  BEGIN
+    DROP PROCEDURE [dbo].[NotificationQueue_GetReady];
+  END;
+GO
+
+IF OBJECT_ID(N'[dbo].[NotificationQueue]', N'U') IS NOT NULL
+  BEGIN
+    DROP TABLE [dbo].[NotificationQueue];
+  END;
+GO